+++ tridesclous/gui/tests/test_peelerwindow.py
@@ -129,7 +129,7 @@ def test_PeelerWindow():
     
     
 if __name__ == '__main__':
-    #~ setup_module()
+    setup_module()
     
     #~ test_Peelercontroller()
