+++ tridesclous/peeler_tools.py
@@ -35,7 +35,7 @@ def make_prediction_one_spike(spike_index, cluster_idx, spike_jitter, dtype, cat
     
     pos = spike_index + catalogue['n_left']
     #~ if spike_jitter is None or np.isnan(spike_jitter):
-    if not catalogue['inter_sample_oversampling']:
+    if not catalogue['inter_sample_oversampling'] or spike_jitter is None or np.isnan(spike_jitter):
         pred = catalogue['centers0'][cluster_idx, :, :]
     else:
         #TODO debug that sign   >>>> done this is correct
@@ -80,49 +80,5 @@ def make_prediction_signals(spikes, dtype, shape, catalogue, safe=True):
     return prediction
 
 
-def get_auto_params_for_peelers(dataio, chan_grp=0):
-    nb_chan = dataio.nb_channel(chan_grp=chan_grp)
-    params = {}
-    
-    
-    #~ params['chunksize'] = int(dataio.sample_rate * 0.5)
-    params['chunksize'] = int(dataio.sample_rate * 0.1)
-    #~ params['chunksize'] = int(dataio.sample_rate * 0.033)
 
-    #~ params['inter_sample_oversampling'] = True
-    params['inter_sample_oversampling'] = False
-    #~ if dataio.sample_rate < 25000.:
-        #~ params['inter_sample_oversampling'] = True
-    #~ else:
-        #~ params['inter_sample_oversampling'] = False
-    
-    #~ if nb_chan <=8:
-    #~ if nb_chan <=1:
-    if nb_chan <=4:
-        
-        params['use_sparse_template'] = False
-        params['sparse_threshold_mad'] = 1.5
-        params['argmin_method'] = 'numpy'
-        
-        params['engine'] = 'classic'
-        
-    else:
-        params['use_sparse_template'] = True
-        params['sparse_threshold_mad'] = 1.5
-        
-        #~ if HAVE_PYOPENCL:
-            #~ params['argmin_method'] = 'opencl'
-            #~ params['engine'] = 'geometrical'
-            #~ params['engine'] = 'geometrical_opencl' # Still experimental
-        #~ elif HAVE_NUMBA:
-            #~ params['argmin_method'] = 'numba'
-            #~ params['engine'] = 'geometrical'
-        #~ else:
-            #~ params['argmin_method'] = 'numpy'
-            #~ params['engine'] = 'geometrical'
-        
-        params['engine'] = 'geometrical'
-        params['argmin_method'] = 'numba'
-        
-    return params
