+++ tridesclous/pruningshears.py
@@ -138,16 +138,7 @@ class PruningShears:
         t1 = time.perf_counter()
         self.log('explore_split_loop', t1-t0)
         
-        
-        #~ t0 = time.perf_counter()
-        #~ cluster_labels = self.try_oversplit(cluster_labels)
-        #~ t1 = time.perf_counter()
-        #~ self.log('try_oversplit', t1-t0)
-        
-        #~ t0 = time.perf_counter()
-        #~ cluster_labels = self.merge_and_clean(cluster_labels)
-        #~ t1 = time.perf_counter()
-        #~ self.log('merge_and_clean', t1-t0)
+
         
         return cluster_labels
     
@@ -728,7 +719,9 @@ class PruningShears:
             self._plot_debug(actual_chan, ind_l0, channel_features_l0, labels_l0, possible_labels_l0, ind_l2, labels_l2, possible_labels_l2, final_label_l2)
             
         self.log('END loop', np.sum(cluster_labels==-1))
+        
         return cluster_labels
+
 
     def _plot_debug(self, actual_chan, ind_l0, features_l0, labels_l0, possible_labels_l0, candidate_labels_l0, final_label):
         if not self.debug_plot:
@@ -852,11 +845,7 @@ class PruningShears:
                     else:
                         lw = 1
                 ax.axvline(self.width * c - self.n_left, color='m', lw=lw)
-        
-        
-        
-        
-        
+
         ax = axs[0, 2]
         if final_label is not None:
             
@@ -909,280 +898,4 @@ class PruningShears:
         #~ self.debug_plot = False
         
         return
-    
-    
-    def try_oversplit(self, cluster_labels):
-        cluster_labels = cluster_labels.copy()
-        
-        labels = np.unique(cluster_labels)
-        labels = labels[labels>=0]
-        
-        m = np.max(labels) + 1
-        
-        
-        for label in np.unique(labels):
-            #~ print()
-            #~ print('label', label)
-
-            ind_keep,  = np.nonzero(cluster_labels == label)
-            if self.dense_mode:
-                #~ waveforms = self.waveforms.take(ind_keep, axis=0)
-                waveforms = self.cc.get_some_waveforms(self.peak_index[ind_keep], channel_indexes=None)
-                extremum_channel = 0
-                centroid = np.median(waveforms, axis=0)
-            else:
-                #~ centroid = np.median(self.waveforms[ind_keep, :, :], axis=0)
-                waveforms = self.cc.get_some_waveforms(self.peak_index[ind_keep], channel_indexes=None)
-                centroid = np.median(waveforms, axis=0)
-                
-                if self.peak_sign == '-':
-                    extremum_channel = np.argmin(centroid[-self.n_left,:], axis=0)
-                elif self.peak_sign == '+':
-                    extremum_channel = np.argmax(centroid[-self.n_left,:], axis=0)
-                # TODO by sparsity level threhold and not radius
-                #~ adjacency = self.channel_adjacency[extremum_channel]
-                #~ waveforms = self.waveforms.take(ind_keep, axis=0).take(adjacency, axis=2)
-                high_adjacency = self.channel_high_adjacency[extremum_channel]
-                #~ waveforms = self.waveforms.take(ind_keep, axis=0).take(high_adjacency, axis=2)
-                waveforms = waveforms.take(high_adjacency, axis=2)
-                
-                
-            wf_flat = waveforms.swapaxes(1,2).reshape(waveforms.shape[0], -1)
-
-            #~ pca =  sklearn.decomposition.IncrementalPCA(n_components=self.n_components_local_pca, whiten=True)
-            
-            n_components = min(wf_flat.shape[1]-1, self.n_components_local_pca)
-            pca =  sklearn.decomposition.TruncatedSVD(n_components=n_components)
-            
-            
-            feats = pca.fit_transform(wf_flat)
-            pval = diptest(np.sort(feats[:, 0]), numt=200)
-            #~ print('pval', pval)
-            self.log('label', label,'pval', pval)
-            #~ if pval<0.2:
-            if pval<0.1: # TODO global params
-            
-                clusterer = hdbscan.HDBSCAN(min_cluster_size=self.min_cluster_size, allow_single_cluster=False, metric='l2')
-                sub_labels = clusterer.fit_predict(feats[:, :2])
-                unique_sub_labels = np.unique(sub_labels)
-                if unique_sub_labels.size ==  1 and unique_sub_labels[0] == -1:
-                    unique_sub_labels = []
-                    self.centroids[label] = centroid
-                    
-                else:
-                    possible_labels, candidate_mask, elsewhere_mask, best_chan_peak_values, best_chan,\
-                        peak_is_aligned, peak_is_on_chan, local_centroids = self.check_candidate_labels(ind_keep, sub_labels, extremum_channel)
-                    
-                    self.log('possible_labels',possible_labels)
-                    
-                    # TODO to avoid split then merge check equal_template
-                    #~ n_shift = 2
-                    #~ for i in range(unique_sub_labels.size):
-                        #~ k1 = unique_sub_labels[i]
-                        #~ for j in range(i+1, unique_sub_labels.size):
-                            #~ k2 = unique_sub_labels[j]
-                            
-                            #~ thresh = max(extremum_amplitude[k1], extremum_amplitude[k2]) * amplitude_factor_thresh
-                            #~ thresh = max(thresh, self.auto_merge_threshold)
-                            #~ if equal_template(local_centroids[k1], local_centroids[k2], thresh=thresh, n_shift=n_shift):
-                    
-                    valid_sub_labels = possible_labels[peak_is_aligned]
-                    
-                    for sub_label in unique_sub_labels:
-                        sub_mask = sub_labels == sub_label
-                        
-                        valid = sub_label in valid_sub_labels
-                        #~ print('sub_label', 'valid', valid)
-                        
-                        if sub_label == -1 or not valid:
-                            cluster_labels[ind_keep[sub_mask]] = -1
-                        else:
-                            cluster_labels[ind_keep[sub_mask]] = sub_label + m 
-                    
-                    if np.max(unique_sub_labels) >=0:
-                        m += np.max(unique_sub_labels) + 1
-                
-                #~ self.debug_plot = True
-                if self.debug_plot:
-                #~ if True:
-                    fig, axs = plt.subplots(ncols=4)
-                    colors = plt.cm.get_cmap('Set3', len(unique_sub_labels))
-                    colors = {unique_sub_labels[l]:colors(l) for l in range(len(unique_sub_labels))}
-                    colors[-1] = 'k'
-                    
-                    for sub_label in unique_sub_labels:
-                        valid = sub_label in possible_labels[peak_is_aligned]
-                        sub_mask = sub_labels == sub_label
-                        if valid:
-                            ls = '-'
-                            color = colors[sub_label]
-                        else:
-                            ls = '--'
-                            color = 'k'
-
-                        ax = axs[0]
-                        ax.plot(wf_flat[sub_mask].T, color=color, alpha=0.1)
-                            
-                        ax = axs[3]
-                        if sub_label>=0:
-                            ax.plot(np.median(wf_flat[sub_mask], axis=0), color=color, lw=2, ls=ls)
-                    
-                    for sub_label in unique_sub_labels:
-                        valid = sub_label in possible_labels[peak_is_aligned]
-                        sub_mask = sub_labels == sub_label
-                        color = colors[sub_label]
-                        if valid:
-                            color = colors[sub_label]
-                        else:
-                            color = 'k'
-                        ax = axs[1]
-                        ax.plot(feats[sub_mask].T, color=color, alpha=0.1)
-                    
-                        ax = axs[2]
-                        ax.scatter(feats[sub_mask][:, 0], feats[sub_mask][:, 1], color=color)
-                    plt.show()
-
-            else:
-                # Not over spplit
-                self.centroids[label] = centroid
-                
-        return cluster_labels
-
-
-
-    def merge_and_clean(self, cluster_labels):
-        # merge : 
-        #   * max distance < auto_merge_threshold
-        #   * 2 cluster have a shift
-        # delete:
-        #   * peak is not aligneds
-        
-        cluster_labels2 = cluster_labels.copy()
-        
-        
-        # self.centroids already computed before
-        
-        extremum_channel = {}
-        extremum_index = {}
-        extremum_amplitude = {}
-        
-        while True:
-            self.log('')
-            self.log('new loop')
-            labels = np.unique(cluster_labels2)
-            labels = labels[labels>=0]
-            n = labels.size
-            self.log(labels)
-            
-            for ind, k in enumerate(labels):
-                if k not in self.centroids:
-                    ind_keep,  = np.nonzero(cluster_labels2 == k)
-                    if ind_keep.size > self.max_per_cluster_for_median:
-                        sub_sel = np.random.choice(ind_keep.size, self.max_per_cluster_for_median, replace=False)
-                        ind_keep = ind_keep[sub_sel]
-                    
-                    waveforms = self.cc.get_some_waveforms(self.peak_index[ind_keep], channel_indexes=None)
-                    centroid = np.median(waveforms, axis=0)
-                    self.centroids[k] = centroid
-                
-                if k not in extremum_index:
-                    centroid = self.centroids[k]
-                    
-                    if self.peak_sign == '-':
-                        chan_peak = np.argmin(np.min(centroid, axis=0))
-                        pos_peak = np.argmin(centroid[:, chan_peak])
-                        peak_val = centroid[-self.n_left, chan_peak]
-                    elif self.peak_sign == '+':
-                        chan_peak = np.argmax(np.max(centroid, axis=0))
-                        pos_peak = np.argmax(centroid[:, chan_peak])
-                        peak_val = centroid[-self.n_left, chan_peak]
-                    extremum_index[k] = pos_peak
-                    extremum_channel[k] = chan_peak
-                    extremum_amplitude[k] = np.abs(peak_val)
-            
-            
-            #eliminate when best peak not aligned
-            # eliminate when peak value is too small
-            # TODO move this in the main loop!!!!!
-            for ind, k in enumerate(labels):
-                
-                if np.abs(-self.n_left - extremum_index[k])>2:
-                    self.log('remove not aligned peak', 'k', k)
-                    #delete
-                    cluster_labels2[cluster_labels2 == k] = -1
-                    labels[ind] = -1
-                    self.centroids.pop(k)
-                    #~ extremum_index.pop(k) # usefull ?
-                    #~ extremum_channel.pop(k) # usefull ?
-                    #~ extremum_amplitude.pop(k) # usefull ?
-                    continue
-                
-                # TODO check if this is really  relevant
-                if np.abs(extremum_amplitude[k]) < self.threshold + 0.5:
-                    self.log('remove small peak', 'k', k, 'peak_val', extremum_amplitude[k])
-                    cluster_labels2[cluster_labels2 == k] = -1
-                    labels[ind] = -1
-                    self.centroids.pop(k)
-                    #~ extremum_index.pop(k) # usefull ?
-                    #~ extremum_channel.pop(k) # usefull ?
-                    #~ extremum_amplitude.pop(k) # usefull ?
-
-                    if self.debug_plot:
-                        fig, ax = plt.subplots()
-                        ax.plot(centroid.T.flatten())
-                        for i in range(self.centroids[k].shape[1]):
-                            ax.axvline(i*self.width-self.n_left)
-                        ax.set_title('delete')
-                        plt.show()
-                    
-                    continue
-                
-            
-            n_shift = 2
-            nb_merge = 0
-            amplitude_factor_thresh = 0.2
-            
-            pop_from_centroids = []
-            for i in range(n):
-                k1 = labels[i]
-                if k1 == -1:
-                    # this can have been removed yet
-                    continue
-                
-                for j in range(i+1, n):
-                    k2 = labels[j]
-                    if k2 == -1:
-                        # this can have been removed yet
-                        continue
-                    
-                    thresh = max(extremum_amplitude[k1], extremum_amplitude[k2]) * amplitude_factor_thresh
-                    thresh = max(thresh, self.auto_merge_threshold)
-                    #~ print('thresh', thresh)
-                    
-                    if equal_template(self.centroids[k1], self.centroids[k2], thresh=thresh, n_shift=n_shift):
-                        self.log('merge', k1, k2)
-                        cluster_labels2[cluster_labels2==k2] = k1
-                        nb_merge += 1
-                        
-                        # remove from centroid doct to recompute it
-                        pop_from_centroids.append(k1)
-                        pop_from_centroids.append(k2)
-                        
-                        if self.debug_plot:
-                            fig, ax = plt.subplots()
-                            ax.plot(self.centroids[i].T.flatten())
-                            ax.plot(self.centroids[j].T.flatten())
-                            ax.set_title('merge')
-                            plt.show()
-            
-            for k in np.unique(pop_from_centroids):
-                if k in self.centroids:
-                    self.centroids.pop(k)
-            
-            self.log('nb_merge', nb_merge)
-            if nb_merge == 0:
-                break
-        
-        return cluster_labels2
-
